chore: moving ai/* to ai/roadmap
